chore: clean up natspec

+++ src/tokens/extensions/RoyaltyManager.sol
@@ -2,9 +2,9 @@
 pragma solidity 0.8.23;
 
 import {IRoyaltyManager, RoyaltyInfo} from "src/interfaces/IRoyaltyManager.sol";
-import {FEE_DENOMINATOR, MAX_ROYALTY_BPS, ALLOCATION_DENOMINATOR} from "src/utils/Constants.sol";
-import {SPLITS_MAIN} from "script/utils/Constants.sol";
 import {ISplitsMain} from "src/interfaces/ISplitsMain.sol";
+
+import "src/utils/Constants.sol";
 
 /**
  * @title RoyaltyManager
@@ -37,7 +37,7 @@ abstract contract RoyaltyManager is IRoyaltyManager {
         uint256 _tokenId
     ) external view returns (address[] memory receivers, uint256[] memory basisPoints) {
         RoyaltyInfo storage tokenRoyalties_ = tokenRoyalties[_tokenId];
-        if (tokenRoyalties_.receiver != address(0) && tokenRoyalties_.basisPoints != uint96(0)) {
+        if (tokenRoyalties_.receiver != address(0) && tokenRoyalties_.basisPoints != 0) {
             receivers = new address[](2);
             basisPoints = new uint256[](2);
             receivers[1] = tokenRoyalties_.receiver;
@@ -58,9 +58,9 @@ abstract contract RoyaltyManager is IRoyaltyManager {
         uint256 _salePrice
     ) external view returns (address receiver, uint256 amount) {
         RoyaltyInfo storage tokenRoyalties_ = tokenRoyalties[_tokenId];
-        if (tokenRoyalties_.receiver != address(0) && tokenRoyalties_.basisPoints != uint96(0)) {
+        if (tokenRoyalties_.receiver != address(0) && tokenRoyalties_.basisPoints != 0) {
             revert MoreThanOneRoyaltyReceiver();
-        } else if (baseRoyalties.receiver == address(0) && baseRoyalties.basisPoints == uint96(0)) {
+        } else if (baseRoyalties.receiver == address(0) && baseRoyalties.basisPoints == 0) {
             return (receiver, amount);
         } else {
             receiver = baseRoyalties.receiver;
@@ -75,8 +75,8 @@ abstract contract RoyaltyManager is IRoyaltyManager {
     /**
      * @notice Sets the base royalties for all tokens
      * @param _receivers Array of addresses receiving royalties
-     * @param _allocations Array of allocations used to calculate royalty payments
-     * @param _basisPoints basis points used to calculate the royalty payment
+     * @param _allocations Array of allocation amounts for calculating royalty shares
+     * @param _basisPoints Total allocation scalar for calculating royalty shares
      */
     function _setBaseRoyalties(
         address[] calldata _receivers,
@@ -100,8 +100,8 @@ abstract contract RoyaltyManager is IRoyaltyManager {
     /**
      * @notice Sets the royalties for a specific token ID
      * @param _tokenId ID of the token
-     * @param _receiver the address receiving royalty payments
-     * @param _basisPoints Array of points used to calculate royalty payments (0.01% per receiver)
+     * @param _receiver Address receiving royalty payments
+     * @param _basisPoints Total allocation scalar for calculating royalty shares
      */
     function _setTokenRoyalties(uint256 _tokenId, address _receiver, uint96 _basisPoints) internal {
         if (!_exists(_tokenId)) revert NonExistentToken();
@@ -118,8 +118,8 @@ abstract contract RoyaltyManager is IRoyaltyManager {
 
     /**
      * @dev Checks if:
-     *        1. the total basis points of royalties exceeds 10,000 (100%)
-     *        2. A single receiver exceeds 25,000 (25%)
+     * 1. Total basis points of royalties exceeds 10,000 (100%)
+     * 2. A single receiver exceeds 2,500 (25%)
      */
     function _checkRoyalties(
         address[] memory _receivers,
